requirements.txt
pandas
numpy
pyspellchecker
nltk
flask